+++ libs/containers/src/yaml.cpp
@@ -7,7 +7,7 @@
    | Released under BSD License. See: https://www.mrpt.org/License          |
    +------------------------------------------------------------------------+ */
 
-#include "containers-precomp.h"	 // Precompiled headers
+#include "containers-precomp.h"  // Precompiled headers
 //
 #include <mrpt/config.h>
 #include <mrpt/containers/yaml.h>
@@ -729,9 +729,9 @@ bool yaml::internalPrintAsYAML(
 			return true;
 	}
 	else if (v.type() == typeid(float))
-		o << mrpt::format("%g", std::any_cast<float>(v));
+		o << mrpt::format("%.16g", std::any_cast<float>(v));
 	else if (v.type() == typeid(double))
-		o << mrpt::format("%g", std::any_cast<double>(v));
+		o << mrpt::format("%.16g", std::any_cast<double>(v));
 	else if (v.type() == typeid(uint16_t))
 		o << std::any_cast<uint16_t>(v);
 	else if (v.type() == typeid(int16_t))
@@ -834,42 +834,42 @@ static std::optional<yaml::node_t> recursiveParse(struct fy_parser* p)
 		case FYET_NONE:
 		{
 			PARSER_DBG_OUT("Event: None");
-			fy_parser_event_free(p, event);	 // free event
+			fy_parser_event_free(p, event);  // free event
 			return recursiveParse(p);  // Keep going
 		}
 		break;
 		case FYET_STREAM_START:
 		{
 			PARSER_DBG_OUT("Event: Stream start");
-			fy_parser_event_free(p, event);	 // free event
+			fy_parser_event_free(p, event);  // free event
 			return recursiveParse(p);  // Keep going
 		}
 		break;
 		case FYET_STREAM_END:
 		{
 			PARSER_DBG_OUT("Event: Stream end");
-			fy_parser_event_free(p, event);	 // free event
+			fy_parser_event_free(p, event);  // free event
 			return recursiveParse(p);  // Keep going
 		}
 		break;
 		case FYET_DOCUMENT_START:
 		{
 			PARSER_DBG_OUT("Event: Doc start");
-			fy_parser_event_free(p, event);	 // free event
+			fy_parser_event_free(p, event);  // free event
 			return recursiveParse(p);  // Keep going
 		}
 		break;
 		case FYET_DOCUMENT_END:
 		{
 			PARSER_DBG_OUT("Event: Doc end");
-			fy_parser_event_free(p, event);	 // free event
+			fy_parser_event_free(p, event);  // free event
 			return recursiveParse(p);  // Keep going
 		}
 		break;
 		case FYET_MAPPING_START:
 		{
 			PARSER_DBG_OUT("Event: MAP START");
-			fy_parser_event_free(p, event);	 // free event
+			fy_parser_event_free(p, event);  // free event
 
 			yaml::node_t n;
 			yaml::map_t& m = n.d.emplace<yaml::map_t>();
@@ -898,21 +898,21 @@ static std::optional<yaml::node_t> recursiveParse(struct fy_parser* p)
 		case FYET_MAPPING_END:
 		{
 			PARSER_DBG_OUT("Event: MAP END");
-			fy_parser_event_free(p, event);	 // free event
+			fy_parser_event_free(p, event);  // free event
 			return {};
 		}
 		break;
 		case FYET_SEQUENCE_START:
 		{
 			PARSER_DBG_OUT("Event: SEQ START");
-			fy_parser_event_free(p, event);	 // free event
+			fy_parser_event_free(p, event);  // free event
 			yaml::node_t n;
 			yaml::sequence_t& s = n.d.emplace<yaml::sequence_t>();
 
 			for (;;)
 			{
 				auto entry = recursiveParse(p);
-				if (!entry.has_value()) break;	// end of sequence reached?
+				if (!entry.has_value()) break;  // end of sequence reached?
 				s.push_back(std::move(entry.value()));
 			}
 
@@ -922,7 +922,7 @@ static std::optional<yaml::node_t> recursiveParse(struct fy_parser* p)
 		case FYET_SEQUENCE_END:
 		{
 			PARSER_DBG_OUT("Event: SEQ END");
-			fy_parser_event_free(p, event);	 // free event
+			fy_parser_event_free(p, event);  // free event
 			return {};
 		}
 		break;
@@ -959,7 +959,7 @@ static std::optional<yaml::node_t> recursiveParse(struct fy_parser* p)
 							CommentPosition::RIGHT)] = std::move(cR.value());
 				}
 
-				fy_parser_event_free(p, event);	 // free event
+				fy_parser_event_free(p, event);  // free event
 				return n;
 			}
 			else
@@ -969,7 +969,7 @@ static std::optional<yaml::node_t> recursiveParse(struct fy_parser* p)
 		}
 		break;
 		case FYET_ALIAS:
-			fy_parser_event_free(p, event);	 // free event
+			fy_parser_event_free(p, event);  // free event
 			return recursiveParse(p);  // Keep going
 	};